Extract the session token decoding to keep blob_store within its budget
`check:architecture` caps `blob_store.rs` at 1000 lines and the new tests took
it to 1005. Five lines over is exactly where trimming comments to fit is most
tempting and least honest, so this extracts instead, as `lifecycle.rs` did for
`diagnostics.rs` in beta.78.

`r2_session_token` holds the decoding and the tests that establish the token's
shape, which belong together: the shape is the entire defect. `blob_store`
keeps a two-line delegation and one test covering the seam, including the case
where no expiry is the truth rather than a failure to read.

blob_store.rs is now 940 lines.
